refacto bug heredoc
