+++ src/ast_tools.c
@@ -26,7 +26,7 @@ t_ast	*create_ast_operator(t_node_type token, t_ast *left, t_ast *right)
 	if (!node)
 		return (NULL);
 	node->token = token;
-	node->cmd = NULL; // change si pipe avant est redi apres
+	node->cmd = NULL;
 	node->left = left;
 	node->right = right;
 	node->head = node;
@@ -56,4 +56,22 @@ int	is_redirection(t_node_type token)
 {
 	return (token == REDIR_IN || token == REDIR_OUT || token == APPEND
 		|| token == REDIR_HEREDOC);
+}
+
+t_redir	*create_redirection(t_node_type type, char *file)
+{
+	t_redir	*redir;
+
+	redir = malloc(sizeof(t_redir));
+	if (!redir)
+		return (NULL);
+	redir->type = type;
+	redir->file = ft_strdup(file);
+	if (!redir->file)
+	{
+		free(redir);
+		return (NULL);
+	}
+	redir->next = NULL;
+	return (redir);
 }
