+++ src/ast_utils.c
@@ -1,4 +1,3 @@
-
 #include "minishell.h"
 #include "pars.h"
