+++ src/signal.c
@@ -6,7 +6,7 @@
 /*   By: pnaessen <pnaessen@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/19 14:11:14 by pnaessen          #+#    #+#             */
-/*   Updated: 2025/02/24 09:46:00 by pnaessen         ###   ########lyon.fr   */
+/*   Updated: 2025/03/12 10:02:07 by pnaessen         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
